Add support for running tests on OSX and Windows
Modify the copy-dependencies plugin configuration to copy dlls, dylibs,
and .so's to the test-lib directory so that tests can be run on more platforms.
Switch the phase to test-compile, since the libs aren't a compilation
requirement.

Update the maven surefire dependency and disable the system classloader for
tests for this bug in some versions of the JDK (like on ubuntu right now):

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=911925