Suggested interface for DynamoDBConfiguration
This is an intermediate PR, not intended for merging but just to get feedback.
It's based on the Python dynamodb crypto config object, which contains
a material provider, encryptioncontext, and attribute actions object.

The idea would be that you call an encryptor like this:

encryptor.encrypt(myRecord, myConfiguration);
encryptor.decrypt(myEncryptedRecord, myConfiguration);

and the myConfiguration object contains everything needed for the job.

The configuration object would be immutable, and the encryptor would contain a
.getConfiguration and .setConfiguration method for controlling defaults.