Add testng and maven failsafe for integration tests
Add a dependency on testng for writing unit and integration tests. Soon, we will
migrate off of junit and onto testng. I've actually already made that change
locally, but it will be less painful to include that with a different cr.

Why maven failsafe? Maven failsafe is like maven surefire, but it ensures that
various cleanup is called when integration tests fail. It hooks into the
existing release workflow, so integration tests run before install or release.