Remove parent dependency from sdk1, sdk2
1.14.0 fails since
software.amazon.cryptools:dynamodbencryptionclient-pom isn't deployed to
maven (but it passes locally, if you've installed it into your local
maven repository).

For context, we originally split the maven module up into a shared
package, an sdk1 package and an sdk2 package, and intended to have a
large amount of reuse. That wound up not happening: sdk2 is entirely
self-contained.

Although there's some duplication in between the build logic between
sdk1 and sdk2, it's not worth the mental overhead of dealing with a
parent and aggregator pom.