feat: Adds the CachingMostRecentProvider. Deprecates MostRecentProvider.
Time-based key re-authorization logic in MostRecentProvider did not re-authorize the use of the key
after key usage permissions were changed at the key provider (for example AWS Key Management
Service). This created the potential for keys to be used in the DynamoDB Encryption Client after
permissions to do so were revoked.

The MostRecentProvider is deprecated. It is removed in 2.0.0. New deployments should use the
CachingMostRecentProvider, and existing deployments should upgrade as soon as possible. See
https://docs.aws.amazon.com/dynamodb-encryption-client/latest/devguide/most-recent-provider.html#mrp-versions
for more details.

This change also addresses interoperability issues between the Python and Java implementations of
the DynamoDB Encryption Client.